+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 1.2
 ---
 
+ * Added `getCustomDateFormat()` for post entities to be able to turn their timestamps into any format
  * Added additional "Go home" item to reset category filters
  * Since now, the active category name is highlighted as green when browsing the grid
  * Added ability to fetch mostly viewed posts. Now users can simply call `getMostlyViewed()` on `$blog` service to get a collection of entities